+++ src/main/java/pokemones/tipoAire.java
@@ -7,7 +7,7 @@ public class tipoAire extends Mounstritos {
 
     private ListaDE<String> debilidades;
 
-    public tipoAire(String name, int HP,ListaDE<Movimientos> ataques, int defense, int ataque, String sprite,ListaDE<String> debilidades) {
+    public tipoAire(String name, float HP,ListaDE<Movimientos> ataques, int defense, int ataque, String sprite,ListaDE<String> debilidades) {
         super(name, HP, ataques, defense, ataque, sprite);
         this.debilidades = debilidades;
     }
