+++ src/main/java/pokemones/tipoRoca.java
@@ -7,7 +7,7 @@ public class tipoRoca extends Mounstritos {
 
     private ListaDE<String> debilidades;
 
-    public tipoRoca(String name, int HP, ListaDE<Movimientos> ataques, int defense, int ataque, String sprite,ListaDE<String> debilidades) {
+    public tipoRoca(String name, float HP, ListaDE<Movimientos> ataques, int defense, int ataque, String sprite,ListaDE<String> debilidades) {
         super(name, HP,ataques, defense, ataque, sprite);
         this.debilidades = debilidades;
     }
